backend/facet/departamentos/apis/carrera.py
from rest_framework import viewsets, filters
from django_filters.rest_framework import DjangoFilterBackend
from ..models import Carrera
from ..serializers import CarreraSerializer

class CarreraViewSet(viewsets.ModelViewSet):
    queryset = Carrera.objects.all()
    serializer_class = CarreraSerializer
    filter_backends = [DjangoFilterBackend, filters.SearchFilter]
    filterset_fields = ['nombre']
    search_fields = ['nombre']