backend/facet/requirements.txt
# Django
Django==4.2.1
# Manejo de variables de entorno
django-environ==0.10.0
# Separa las variables de conexion en un archivo .venv
python-dotenv==1.0.0
python-dateutil==2.8.2
# Sacar Acento
unidecode==1.3.8
# Simple History
django-simple-history==3.3.0
# Tema para django-admin
django-jazzmin==2.6.0
# Postgresql
psycopg2==2.9.6
# Encriptacion de clave
argon2-cffi==21.3.0
# Django REST framework
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-rest-knox==4.2.0
# Encabezados de uso compartido de recursos de origen cruzado 
django-cors-headers==3.13.0
#Convierte HTML a texto
bs4==0.0.2
# Swagger
drf-yasg==1.21.5